Initial re-org for 2.0 compatibility
