--- MITIE/README.md
@@ -1,5 +1,0 @@
-# MITIE folder
-
-This is where the feature extractor needs to be placed (see instructions under **Install** in main README.md)
-
-The files are not included as part of this repo (as they are large and hosted elsewhere as part of the MITIE project)
