+++ config/README.md
@@ -2,6 +2,4 @@
 
 This is where the various configuration files are stored.
 
-Some of the details will depend on the model you use and others will depend on the email account you use (should you choose to do so, it is optional and not necessary to use the bot unless you particularly wish to use it over email)
-
 Please take care not to accidentally post your own credentials if you fork the project!
