--- config/config.json
@@ -1,7 +1,0 @@
-{
-  "backend": "mitie_sklearn",
-  "mitie_file" : "./MITIE/MITIE-models/english/total_word_feature_extractor.dat",
-  "path" : "./models/",
-  "data" : "./data/RoyBot.json",
-  "server_model_dir":"./model_20170228-111802"
-}
