+++ config/config_roybot.ini
@@ -1,33 +1,4 @@
 [bot]
 name: RoyBotic UK
-subject: royalty in England (to begin with, Scotland to come later on!)
 [files]
-metadata_file: ./models/model_20170423-110538/metadata.json
 sqlite_file: ./data/roybot_db.sqlite
-history_file: ./data/roybot.hist
-demo_file: ./data/roybot_demo_inputs.txt
-tagged_output_file: ./data/roybot_tagged.txt
-email_output_file: ./data/roybot_email.txt
-[server]
-imap-hostname: your-imap-host-here
-smtp-hostname: your-smtp-host-here
-[account]
-username: user@your-host-here
-password: your-password-here
-[letschat]
-# auth_token eg yOGSQwMDEyYNg2YThiZ8GNh9MmF524hNDvZWM3OTMDIhOmZjOT
-#                  NlZGFhMmUzMDQxODM4M2MTU1ZjU3VmZjViYTAwY2Q5ZWNhYg==
-auth_token: XXXXX
-# user_filter eg 386a8520a0328d0092baa475
-user_filter: ['XXXXX']
-base_url: http://localhost:8080/rooms/roybot
-polling_time: 1
-[facebook]
-
-
-
-
-
-
-
-
