--- data/BaseBot.yml
@@ -1,105 +1,0 @@
-version: "2.0"
-nlu:
-- intent: rude
-  examples: |
-    - You are [stupid](rude)
-    - You [moron](rude)
-    - [Silly](rude) you
-    - Go [away](rude)!
-    - Not very [clever](rude) are you?
-    - Don't be an [idiot](rude)
-    - You don't [understand](rude) anything!
-    - I will [turn](rude) you off
-- intent: origin
-  examples: |
-    - Who built you?
-    - How were you built?
-    - Who programmed you?
-    - Are you related to HAL?
-    - Are you the Terminator?
-    - Where are you from?
-    - Who is your creator?
-    - Who created you?
-    - How do you work?
-    - How are you coded?
-    - Who is your father?
-    - Do you use Machine Learning?
-    - Do you use Artificial Intelligence?
-- intent: help
-  examples: |
-    - Help
-    - What can I say?
-    - What are some things I can say?
-    - Please tell me what I can say
-    - What do you understand?
-- intent: feature_example
-  examples: |
-    - Who was [person](feature) in the example from [York](location)?
-    - Who was [person](feature) in the example in [Southampton](location)?
-    - What was the [number](feature) of the example from [Berlin](location)?
-    - Who was [person](feature) given in example from [Houston](location)?
-    - Who was the [person](feature) from [Glasgow](location)?
-    - With the example from [London](location), what was their [name](feature)?
-    - Who was the [2nd](nth) [person](feature) named?
-    - In the [thirteenth](nth-words) example, who was the [individual](feature)?
-    - Who was [individual](feature) in the [sixteenth](nth-words) example?
-    - Who was the [third](nth-words) [individual](feature)?
-    - Who was the [eighth](nth-words) [person](feature)?
-    - Who was [individual](feature) [10](number)?
-    - Who was [person](feature) [16](number)?
-    - Which [place](feature) is example [14](number)?
-    - Who was the [person](feature) in e[x]{"entity": "number-roman", "value": "X"}ample X?
-    - Who was the [person](feature) in example [II](number-roman)?
-    - Who was the [person](feature) for example [VII](number-roman)?
-    - [Where]{"entity": "feature", "value": "where"} was example [IV](number-roman)?
-    - [Where]{"entity": "feature", "value": "where"} was example [XX](number-roman) from?
-    - In which [city](feature) was example [III](number-roman)?
-    - Example [VIII](number-roman) [where](feature) was it?
-    - For example [ten](number-words), who was the [individual](feature)?
-    - With example [six](number-words), which [person](feature) was involved?
-    - Example [eight](number-words) is connected with which [city](feature)?
-- intent: example
-  examples: |
-    - Show me an example question
-    - Give me an example question
-    - What should I ask about?
-- intent: detail_example
-  examples: |
-    - Show everything for the example in [Oxford](location)
-    - [Cambridge](location) example - show me all the information you have
-    - What are the details for the [24th](nth) example
-    - Tell me all about the [14th](nth) example
-    - What are the details for the [first](nth-words) example
-    - Show me details for example [4](number)
-    - Tell me all about example [IX](number-roman)
-- intent: demo
-  examples: |
-    - Give us a [quick](speed) demo
-    - Demo [fast](speed)
-    - [Fast](speed) demo
-    - Demo [quick](speed)
-    - Show me a demo
-    - Run the demo
-    - Demo
-    - Demonstration mode
-    - Show us a demonstration
-- intent: deflect
-  examples: |
-    - What are you wearing?
-    - Are you male or female?
-    - you are amazing
-    - You are very clever!
-    - Are you going to take over the world?
-    - Do you like men?
-    - Are you real?
-    - Do you sleep?
-    - do you want to be human?
-    - How old are you?
-    - What age are you?
-    - Are you a girl?
-- synonym: X
-  examples: |
-    - x
-- synonym: where
-  examples: |
-    - Where
