--- data/base_demo_inputs.txt
@@ -1,3 +1,0 @@
-Who made you?
-Help
-What can I say?
