--- data/create_base_example_table.sql
@@ -1,2 +1,0 @@
---
-CREATE TABLE `example` ( `Number` , `Name` , `Location`)
