+++ requirements.txt
@@ -1,18 +1,5 @@
-arrow==0.10.0
-colored==1.3.3
-imaplib2==2.45.0
-numpy==1.11.3
-rasa-nlu==0.5.0.0
-requests==2.20.0
-roman==2.0.0
-scipy==0.18.1
-scikit-learn==0.18.1
-spacy==1.4.0
-Flask==1.0
-Jinja2==2.8
-MarkupSafe==0.23
-Werkzeug==0.11.10
-click==6.6
-gunicorn==19.6.0
-itsdangerous==0.24
-wsgiref==0.1.2
+arrow==1.0.3
+requests==2.25
+roman==3.3
+text2digits==0.1.0
+num2words==0.5
